pasm fdecl support + ability to step through assembler with debugger
